pyproject.toml
[project]
name = "blobby"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.0",
    "dotenv>=0.9.9",
    "pip-system-certs>=5.2",
]
